+++ Team_10Game/Team_10Game/Collision.h
@@ -8,8 +8,8 @@ class Collision
 {
 public:
 	//円と円の衝突判定
-	static bool CircleToCircle(XMFLOAT3 centerPosA, int radiusA,
-		XMFLOAT3 centerPosB, int radiusB);
+	static bool CircleToCircle(XMFLOAT3 centerPosA, float radiusA,
+		XMFLOAT3 centerPosB, float radiusB);
 
 	//円と平面
 	//distanceは原点からの距離であり、上に行ったら+、下に行ったら-になる
